Moved SimCrypto into simutils package as well
